lembas-a-service/lembas-core/pyproject.toml
[project]
name = "lembas-core"
version = "0.1.0"
description = "Core library for Lembas Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic",
    "loguru",
    "click>=8.3.0",
    "python-dotenv>=1.2.1",
    "peewee>=3.17.0",
    "psycopg2-binary>=2.9.10",
]

[project.scripts]
lembas-core = "lembas_core.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
